+++ tianshou/policy/imitation/discrete_cql.py
@@ -37,10 +37,6 @@ class DiscreteCQLPolicy(QRDQN):
         you do not use the target network).
     :param reward_normalization: normalize the **returns** to Normal(0, 1).
         TODO: rename to return_normalization?
-    :param is_double: use double dqn.
-    :param clip_loss_grad: clip the gradient of the loss in accordance
-        with nature14236; this amounts to using the Huber loss instead of
-        the MSE loss.
     :param observation_space: Env's observation space.
     :param lr_scheduler: if not None, will be called in `policy.update()`.
 
@@ -61,8 +57,6 @@ class DiscreteCQLPolicy(QRDQN):
         estimation_step: int = 1,
         target_update_freq: int = 0,
         reward_normalization: bool = False,
-        is_double: bool = True,
-        clip_loss_grad: bool = False,
         observation_space: gym.Space | None = None,
         lr_scheduler: TLearningRateScheduler | None = None,
     ) -> None:
@@ -75,8 +69,6 @@ class DiscreteCQLPolicy(QRDQN):
             estimation_step=estimation_step,
             target_update_freq=target_update_freq,
             reward_normalization=reward_normalization,
-            is_double=is_double,
-            clip_loss_grad=clip_loss_grad,
             observation_space=observation_space,
             lr_scheduler=lr_scheduler,
         )
@@ -88,8 +80,7 @@ class DiscreteCQLPolicy(QRDQN):
         *args: Any,
         **kwargs: Any,
     ) -> TDiscreteCQLTrainingStats:
-        if self._target and self._iter % self.freq == 0:
-            self._update_lagged_network_weights()
+        self._periodically_update_lagged_network_weights()
         self.optim.zero_grad()
         weight = batch.pop("weight", 1.0)
         all_dist = self(batch).logits
@@ -115,7 +106,6 @@ class DiscreteCQLPolicy(QRDQN):
         loss = qr_loss + min_q_loss * self.min_q_weight
         loss.backward()
         self.optim.step()
-        self._iter += 1
 
         return DiscreteCQLTrainingStats(  # type: ignore[return-value]
             loss=loss.item(),