+++ tianshou/policy/modelfree/c51.py
@@ -7,9 +7,12 @@ import torch
 
 from tianshou.data import Batch, ReplayBuffer
 from tianshou.data.types import RolloutBatchProtocol
-from tianshou.policy import DQN
 from tianshou.policy.base import TLearningRateScheduler
-from tianshou.policy.modelfree.dqn import DQNPolicy, DQNTrainingStats
+from tianshou.policy.modelfree.dqn import (
+    DQNPolicy,
+    DQNTrainingStats,
+    QLearningOffPolicyAlgorithm,
+)
 from tianshou.utils.net.common import Net
 
 
@@ -57,28 +60,8 @@ class C51Policy(DQNPolicy):
         return super().compute_q_value((logits * self.support).sum(2), mask)
 
 
-class C51(DQN[C51Policy, TC51TrainingStats], Generic[TC51TrainingStats]):
-    """Implementation of Categorical Deep Q-Network. arXiv:1707.06887.
-    :param policy: a policy following the rules (s -> action_values_BA)
-    :param optim: a torch.optim for optimizing the policy.
-    :param discount_factor: in [0, 1].
-    :param estimation_step: the number of steps to look ahead.
-    :param target_update_freq: the target network update frequency (0 if
-        you do not use the target network).
-    :param reward_normalization: normalize the **returns** to Normal(0, 1).
-        TODO: rename to return_normalization?
-    :param is_double: use double dqn.
-    :param clip_loss_grad: clip the gradient of the loss in accordance
-        with nature14236; this amounts to using the Huber loss instead of
-        the MSE loss.
-    :param observation_space: Env's observation space.
-    :param lr_scheduler: if not None, will be called in `policy.update()`.
-
-    .. seealso::
-
-        Please refer to :class:`~tianshou.policy.DQNPolicy` for more detailed
-        explanation.
-    """
+class C51(QLearningOffPolicyAlgorithm[C51Policy, TC51TrainingStats], Generic[TC51TrainingStats]):
+    """Implementation of Categorical Deep Q-Network. arXiv:1707.06887."""
 
     def __init__(
         self,
@@ -89,10 +72,23 @@ class C51(DQN[C51Policy, TC51TrainingStats], Generic[TC51TrainingStats]):
         estimation_step: int = 1,
         target_update_freq: int = 0,
         reward_normalization: bool = False,
-        is_double: bool = True,
-        clip_loss_grad: bool = False,
         lr_scheduler: TLearningRateScheduler | None = None,
     ) -> None:
+        """
+        :param policy: a policy following the rules (s -> action_values_BA)
+        :param optim: a torch.optim for optimizing the policy.
+        :param discount_factor: in [0, 1].
+        :param estimation_step: the number of steps to look ahead.
+        :param target_update_freq: the target network update frequency (0 if
+            you do not use the target network).
+        :param reward_normalization: normalize the **returns** to Normal(0, 1).
+            TODO: rename to return_normalization?
+        :param is_double: use double dqn.
+        :param clip_loss_grad: clip the gradient of the loss in accordance
+            with nature14236; this amounts to using the Huber loss instead of
+            the MSE loss.
+        :param lr_scheduler: if not None, will be called in `policy.update()`.
+        """
         super().__init__(
             policy=policy,
             optim=optim,
@@ -100,8 +96,6 @@ class C51(DQN[C51Policy, TC51TrainingStats], Generic[TC51TrainingStats]):
             estimation_step=estimation_step,
             target_update_freq=target_update_freq,
             reward_normalization=reward_normalization,
-            is_double=is_double,
-            clip_loss_grad=clip_loss_grad,
             lr_scheduler=lr_scheduler,
         )
         self.delta_z = (policy.v_max - policy.v_min) / (policy.num_atoms - 1)
@@ -111,7 +105,7 @@ class C51(DQN[C51Policy, TC51TrainingStats], Generic[TC51TrainingStats]):
 
     def _target_dist(self, batch: RolloutBatchProtocol) -> torch.Tensor:
         obs_next_batch = Batch(obs=batch.obs_next, info=[None] * len(batch))
-        if self._target:
+        if self.use_target_network:
             act = self.policy(obs_next_batch).act
             next_dist = self.policy(obs_next_batch, model=self.model_old).logits
         else:
@@ -135,8 +129,7 @@ class C51(DQN[C51Policy, TC51TrainingStats], Generic[TC51TrainingStats]):
         *args: Any,
         **kwargs: Any,
     ) -> TC51TrainingStats:
-        if self._target and self._iter % self.freq == 0:
-            self._update_lagged_network_weights()
+        self._periodically_update_lagged_network_weights()
         self.optim.zero_grad()
         with torch.no_grad():
             target_dist = self._target_dist(batch)
@@ -150,6 +143,5 @@ class C51(DQN[C51Policy, TC51TrainingStats], Generic[TC51TrainingStats]):
         batch.weight = cross_entropy.detach()  # prio-buffer
         loss.backward()
         self.optim.step()
-        self._iter += 1
 
         return C51TrainingStats(loss=loss.item())  # type: ignore[return-value]