+++ tianshou/policy/modelfree/fqf.py
@@ -109,8 +109,6 @@ class FQF(QRDQN[FQFPolicy, TFQFTrainingStats]):
         estimation_step: int = 1,
         target_update_freq: int = 0,
         reward_normalization: bool = False,
-        is_double: bool = True,
-        clip_loss_grad: bool = False,
         lr_scheduler: TLearningRateScheduler | None = None,
     ) -> None:
         """
@@ -126,10 +124,6 @@ class FQF(QRDQN[FQFPolicy, TFQFTrainingStats]):
             you do not use the target network).
         :param reward_normalization: normalize the **returns** to Normal(0, 1).
             TODO: rename to return_normalization?
-        :param is_double: use double dqn.
-        :param clip_loss_grad: clip the gradient of the loss in accordance
-            with nature14236; this amounts to using the Huber loss instead of
-            the MSE loss.
         :param observation_space: Env's observation space.
         :param lr_scheduler: if not None, will be called in `policy.update()`.
         """
@@ -141,8 +135,6 @@ class FQF(QRDQN[FQFPolicy, TFQFTrainingStats]):
             estimation_step=estimation_step,
             target_update_freq=target_update_freq,
             reward_normalization=reward_normalization,
-            is_double=is_double,
-            clip_loss_grad=clip_loss_grad,
             lr_scheduler=lr_scheduler,
         )
         self.ent_coef = ent_coef
@@ -153,7 +145,7 @@ class FQF(QRDQN[FQFPolicy, TFQFTrainingStats]):
             obs=buffer[indices].obs_next,
             info=[None] * len(indices),
         )  # obs_next: s_{t+n}
-        if self._target:
+        if self.use_target_network:
             result = self.policy(obs_next_batch)
             act, fractions = result.act, result.fractions
             next_dist = self.policy(
@@ -171,8 +163,7 @@ class FQF(QRDQN[FQFPolicy, TFQFTrainingStats]):
         *args: Any,
         **kwargs: Any,
     ) -> TFQFTrainingStats:
-        if self._target and self._iter % self.freq == 0:
-            self._update_lagged_network_weights()
+        self._periodically_update_lagged_network_weights()
         weight = batch.pop("weight", 1.0)
         out = self.policy(batch)
         curr_dist_orig = out.logits
@@ -227,7 +218,6 @@ class FQF(QRDQN[FQFPolicy, TFQFTrainingStats]):
         self.optim.zero_grad()
         quantile_loss.backward()
         self.optim.step()
-        self._iter += 1
 
         return FQFTrainingStats(  # type: ignore[return-value]
             loss=quantile_loss.item() + fraction_entropy_loss.item(),