+++ tianshou/policy/modelfree/rainbow.py
@@ -50,6 +50,6 @@ class RainbowDQN(C51[TRainbowTrainingStats]):
         **kwargs: Any,
     ) -> TRainbowTrainingStats:
         self._sample_noise(self.policy.model)
-        if self._target and self._sample_noise(self.model_old):
+        if self.use_target_network and self._sample_noise(self.model_old):
             self.model_old.train()  # so that NoisyLinear takes effect
         return super()._update_with_batch(batch, **kwargs)